+++ textext/asktext.py
@@ -2,7 +2,7 @@
 This file is part of TexText, an extension for the vector
 illustration program Inkscape.
 
-Copyright (c) 2006-2024 TexText developers.
+Copyright (c) 2006-2025 TexText developers.
 
 TexText is released under the 3-Clause BSD license. See
 file LICENSE.txt or go to https://github.com/textext/textext
@@ -140,7 +140,7 @@ class AskText(object):
     CLOSE_SHORTCUT = ["Escape", "CtrlQ", "None"]
 
     def __init__(self, version_str, text, preamble_file, global_scale_factor, current_scale_factor, current_alignment,
-                 current_texcmd, current_convert_strokes_to_path, tex_commands, gui_config):
+                 current_texcmd, tex_commands, gui_config):
         self.TEX_COMMANDS = tex_commands
         if len(text) > 0:
             self.text = text
@@ -160,10 +160,22 @@ class AskText(object):
             self.current_texcmd = current_texcmd
         else:
             self.current_texcmd = self.TEX_COMMANDS[0]
-
-        self.current_convert_strokes_to_path = current_convert_strokes_to_path
+        self.using_tex = self.current_texcmd != "typst"
 
         self.preamble_file = preamble_file
+
+        # TexText < 1.10 did not use preamble files for typst and just stored
+        # the default_packages.tex. Hence, we have to correct this here.
+        if not self.using_tex and "default_packages.tex" in self.preamble_file:
+            self.preamble_file = "default_preamble_typst.typ"
+
+        if self.using_tex:
+            self.latex_default_preamble_file = self.preamble_file
+            self.typst_default_preamble_file = "default_preamble_typst.typ"
+        else:
+            self.latex_default_preamble_file = "default_packages.tex"
+            self.typst_default_preamble_file = self.preamble_file
+
         self._preamble_widget = None
         self._scale = None
         self._gui_config = gui_config
@@ -220,9 +232,9 @@ class AskTextTK(AskText):
     """TK GUI for editing TexText objects"""
 
     def __init__(self, version_str, text, preamble_file, global_scale_factor, current_scale_factor, current_alignment,
-                 current_texcmd, current_convert_strokes_to_path, tex_commands, gui_config):
+                 current_texcmd, tex_commands, gui_config):
         super(AskTextTK, self).__init__(version_str, text, preamble_file, global_scale_factor, current_scale_factor,
-                                        current_alignment, current_texcmd, current_convert_strokes_to_path, tex_commands, gui_config)
+                                        current_alignment, current_texcmd, tex_commands, gui_config)
         self._frame = None
         self._scale = None
         self._preamble = None
@@ -273,7 +285,7 @@ class AskTextTK(AskText):
         label.pack(pady=2, padx=5, anchor="w")
         self._preamble = Tk.Entry(box)
         self._preamble.pack(expand=True, fill="x", ipady=4, pady=5, padx=5, side="left", anchor="e")
-        self._preamble.insert(Tk.END, self.preamble_file)
+        self.set_default_preamble()
 
         self._askfilename_button = Tk.Button(box, text="Select...",
                                        command=self.select_preamble_file)
@@ -284,15 +296,6 @@ class AskTextTK(AskText):
         # Frame holding the advanced settings and the tex command
         box2 = Tk.Frame(self._frame, relief="flat")
         box2.pack(fill="x", pady=5, expand=True)
-
-        # Frame box for advanced settings
-        self._convert_strokes_to_path = Tk.BooleanVar()
-        self._convert_strokes_to_path.set(self.current_convert_strokes_to_path)
-        box = Tk.Frame(box2, relief="groove", borderwidth=2)
-        label = Tk.Label(box, text="SVG-output:")
-        label.pack(pady=2, padx=5, anchor="w")
-        Tk.Checkbutton(box, text="No strokes", variable=self._convert_strokes_to_path, onvalue=True, offvalue=False).pack(side="left", expand=False, anchor="w")
-        box.pack(side=Tk.RIGHT, fill="x", pady=5, expand=True)
 
         # Frame box for tex command
         self._tex_command_tk_str = Tk.StringVar()
@@ -419,11 +422,10 @@ class AskTextTK(AskText):
             return
         self.text = self._text_box.get(1.0, Tk.END)
         self.preamble_file = self._preamble.get()
-        self.current_convert_strokes_to_path = self._convert_strokes_to_path.get()
 
         try:
             self.callback(self.text, self.preamble_file, self.global_scale_factor, self._alignment_tk_str.get(),
-                          self._tex_command_tk_str.get(), self.current_convert_strokes_to_path)
+                          self._tex_command_tk_str.get())
         except Exception as error:
             self.show_error_dialog("TexText Error",
                               "Error occurred while converting text from Latex to SVG:",
@@ -438,9 +440,15 @@ class AskTextTK(AskText):
         self._gui_config["word_wrap"] = self._word_wrap_tkval.get()
 
     def on_texcmd_change(self):
-        using_tex = self._tex_command_tk_str.get() != "typst"
-        self._preamble["state"] = Tk.NORMAL if using_tex else Tk.DISABLED
-        self._askfilename_button["state"] = Tk.NORMAL if using_tex else Tk.DISABLED
+        self.using_tex = self._tex_command_tk_str.get() != "typst"
+        self.set_default_preamble()
+
+    def set_default_preamble(self):
+        self._preamble.delete(0, Tk.END)
+        if self.using_tex:
+            self._preamble.insert(Tk.END, self.latex_default_preamble_file)
+        else:
+            self._preamble.insert(Tk.END, self.typst_default_preamble_file)
 
     def reset_scale_factor(self, _=None):
         self._scale.delete(0, "end")
@@ -451,9 +459,13 @@ class AskTextTK(AskText):
         self._scale.insert(0, self.global_scale_factor)
 
     def select_preamble_file(self):
+        if self.using_tex:
+            file_types = (("LaTeX files", "*.tex"), ("all files", "*.*"))
+        else:
+            file_types = (("Typst files", "*.typ"), ("all files", "*.*"))
         file_name = TkFileDialogs.askopenfilename(initialdir=os.path.dirname(self._preamble.get()),
                                                   title="Select preamble file",
-                                                  filetypes=(("LaTeX files", "*.tex"), ("all files", "*.*")))
+                                                  filetypes=file_types)
         if file_name is not None:
             self._preamble.delete(0, Tk.END)
             self._preamble.insert(Tk.END, file_name)
@@ -506,10 +518,9 @@ class AskTextGTKSource(AskText):
     """GTK + Source Highlighting for editing TexText objects"""
 
     def __init__(self, version_str, text, preamble_file, global_scale_factor, current_scale_factor, current_alignment,
-                 current_texcmd, current_convert_strokes_to_path, tex_commands, gui_config):
+                 current_texcmd, tex_commands, gui_config):
         super(AskTextGTKSource, self).__init__(version_str, text, preamble_file, global_scale_factor, current_scale_factor,
-                                               current_alignment, current_texcmd, current_convert_strokes_to_path,
-                                               tex_commands, gui_config)
+                                               current_alignment, current_texcmd, tex_commands, gui_config)
         self._preview = None  # type: Gtk.Image
         self._pixbuf = None  # type: GdkPixbuf
         self.preview_representation = "SCALE"  # type: str
@@ -798,13 +809,10 @@ class AskTextGTKSource(AskText):
 
         self.global_scale_factor = self._scale_adj.get_value()
 
-        self.current_convert_strokes_to_path = self._conv_stroke2path.get_active()
-
         try:
             self.callback(self.text, self.preamble_file, self.global_scale_factor,
                           self.ALIGNMENT_LABELS[self._alignment_combobox.get_active()],
-                          self.TEX_COMMANDS[self._texcmd_cbox.get_active()].lower(),
-                          self.current_convert_strokes_to_path)
+                          self.TEX_COMMANDS[self._texcmd_cbox.get_active()].lower())
         except Exception as error:
             self.show_error_dialog("TexText Error",
                                    "Error occurred while converting text from Latex to SVG:",
@@ -819,10 +827,20 @@ class AskTextGTKSource(AskText):
         self.window_deleted_cb(widget, None, None)
 
     def cb_compiler_changed(self, combo_box):
-        using_tex = self.TEX_COMMANDS[self._texcmd_cbox.get_active()] != "typst"
-        self._preview_button.set_sensitive(using_tex)
-        self._preamble_widget.set_sensitive(using_tex)
-        self._preamble_delete_btn.set_sensitive(using_tex)
+        self.using_tex = self.TEX_COMMANDS[self._texcmd_cbox.get_active()] != "typst"
+        self.set_default_preamble()
+
+    def set_default_preamble(self):
+        if self.using_tex:
+            preamble_file_str = self.latex_default_preamble_file
+        else:
+            preamble_file_str = self.typst_default_preamble_file
+
+        if hasattr(Gtk, 'FileChooserButton'):
+            file_path = os.path.abspath(preamble_file_str)
+            self._preamble_widget.set_filename(file_path)
+        else:
+            self._preamble_widget.set_text(preamble_file_str)
 
     def move_cursor_cb(self, text_buffer, cursoriter, mark, view):
         self.update_position_label(text_buffer, self, view)
@@ -983,14 +1001,18 @@ class AskTextGTKSource(AskText):
         """
         Clear the preamble file setting
         """
-        self.preamble_file = "default_packages.tex"
+        if self.using_tex:
+            self.preamble_file = "default_packages.tex"
+        else:
+            self.preamble_file = "default_preamble_typst.typ"
         self.set_preamble()
 
     def set_preamble(self):
         if hasattr(Gtk, 'FileChooserButton'):
-            self._preamble_widget.set_filename(self.preamble_file)
+            file_path = os.path.abspath(self.preamble_file)
+            self._preamble_widget.set_filename(file_path)
         else:
-            self._preamble_widget.set_text(self.preamble_file)
+            self._preamble_widget.set_text(self._preamble_widget.set_filename(self.preamble_file))
 
     def reset_scale_factor(self, _=None):
         self._scale_adj.set_value(self.current_scale_factor)
@@ -1116,19 +1138,10 @@ class AskTextGTKSource(AskText):
 
         alignment_box.pack_start(self._alignment_combobox, True, True, 2)
 
-        # Advanced settings
-        adv_settings_frame = Gtk.Frame()
-        adv_settings_frame.set_label("SVG output")
-        self._conv_stroke2path = Gtk.CheckButton(label="No strokes")
-        self._conv_stroke2path.set_tooltip_text("Ensures that strokes (lines, e.g. in \\sqrt, \\frac) can be easily \ncolored in Inkscape (Time consuming compilation!)")
-        self._conv_stroke2path.set_active(self.current_convert_strokes_to_path)
-        adv_settings_frame.add(self._conv_stroke2path)
-
         # --- Scale, alignment and advanced settings together in one "line"
         scale_align_hbox = Gtk.HBox(homogeneous=False, spacing=0)
         scale_align_hbox.pack_start(scale_frame, False, False, 5)
         scale_align_hbox.pack_start(alignment_frame, True, True, 5)
-        scale_align_hbox.pack_start(adv_settings_frame, True, True, 5)
 
         # --- TeX code window ---
         # Scrolling Window with Source View inside